Moved consolidated vision code from enhancements into normal Subsystems
