+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Subsystems/Vision.java
@@ -5,6 +5,7 @@ import static org.firstinspires.ftc.robotcore.external.navigation.AxesOrder.XYZ;
 import static org.firstinspires.ftc.robotcore.external.navigation.AxesReference.EXTRINSIC;
 
 import com.qualcomm.robotcore.hardware.HardwareMap;
+import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.firstinspires.ftc.robotcore.external.ClassFactory;
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
@@ -12,18 +13,24 @@ import org.firstinspires.ftc.robotcore.external.matrices.OpenGLMatrix;
 import org.firstinspires.ftc.robotcore.external.matrices.VectorF;
 import org.firstinspires.ftc.robotcore.external.navigation.Orientation;
 import org.firstinspires.ftc.robotcore.external.navigation.VuforiaLocalizer;
+import org.firstinspires.ftc.teamcode.Enhancement.Config.VisionConfig;
+import org.firstinspires.ftc.teamcode.Enhancement.Subsystems.Subsystem;
+import org.firstinspires.ftc.teamcode.Enhancement.Subsystems.Vision.DetectMarkerPipeline;
+import org.firstinspires.ftc.teamcode.Enhancement.Subsystems.Vision.DetectMarkerPipeline.MarkerLocation;
+import org.firstinspires.ftc.teamcode.Util.QuickTelemetry;
 import org.openftc.easyopencv.OpenCvCamera;
 import org.openftc.easyopencv.OpenCvCameraFactory;
+import org.openftc.easyopencv.OpenCvCameraRotation;
 import org.openftc.easyopencv.OpenCvInternalCamera;
 
 /**
  * The Vision Subsystem
  *
- * @see DetectMarkerPipeline
+ * @see org.firstinspires.ftc.teamcode.Enhancement.Subsystems.Vision.DetectMarkerPipeline
  * @see <a href="https://github.com/OpenFTC/EasyOpenCV">EasyOpenCV</a>
  */
 
-public class Vision {
+public class Vision extends Subsystem {
     public static final int CAMERA_WIDTH = 320; // width  of wanted camera resolution
     public static final int CAMERA_HEIGHT = 240; // height of wanted camera resolution
     public static final int HORIZON = 100; // horizon value to tune
@@ -56,27 +63,50 @@ public class Vision {
 
     private int[] viewportContainerIds;
 
-    public Vision(HardwareMap hardwareMap, Robot robot, boolean isBlue) {
-        robot.getOpmode().telemetry.addData("Vision Status", "Vision initializing started");
+    // Move stuff
+    HardwareMap hardwareMap;
+    OpenCvInternalCamera robotCamera;
+    MarkerLocation markerLocation = MarkerLocation.NOT_FOUND;
+    QuickTelemetry quickTelemetry;
+
+    /**
+     * Class instantiation
+     *
+     * @param telemetry   Quick Telemetry
+     * @param hardwareMap the hardware map
+     * @param timer       how much time elapsed
+     * @throws InterruptedException It might happen because the thread is interrupted.
+     */
+    public Vision(QuickTelemetry telemetry, HardwareMap hardwareMap, ElapsedTime timer) {
+        super(telemetry, hardwareMap, timer);
+
+        telemetry.telemetry(3, "Vision Status", "Vision initializing started");
 
         webcamName = hardwareMap.get(WebcamName.class, WEBCAM_NAME);
         int cameraMonitorViewId = hardwareMap.appContext.getResources().getIdentifier("cameraMonitorViewId", "id", hardwareMap.appContext.getPackageName());
         viewportContainerIds = OpenCvCameraFactory.getInstance().splitLayoutForMultipleViewports(cameraMonitorViewId, 2, OpenCvCameraFactory.ViewportSplitMethod.HORIZONTALLY);
 
 
-        robot.getOpmode().telemetry.addData("init Vuforia", "init Vuforia started");
+        telemetry.telemetry(3, "init Vuforia", "init Vuforia started");
         initVuforia();
-        robot.getOpmode().telemetry.addData("init Vuforia", "init Vuforia completed");
+        telemetry.telemetry(2, "init Vuforia", "init Vuforia completed");
 
         OpenCvInternalCamera robotCamera;
 
         robotCamera = OpenCvCameraFactory.getInstance().createInternalCamera(OpenCvInternalCamera.CameraDirection.BACK, cameraMonitorViewId);
-
-        robot.getOpmode().telemetry.addData("Detect Marker", "Detecting Marker");
-        DetectMarker detectMarkerRunnable = new DetectMarker(hardwareMap, robotCamera, robot.getOpmode().telemetry);
+/*
+        telemetry.telemetry(4, "Detect Marker", "Detecting Marker");
+        DetectMarker detectMarkerRunnable = new DetectMarker(hardwareMap, robotCamera, telemetry);
         finalMarkerLocation = detectMarkerRunnable.DetectMarkerRun();
-        robot.getOpmode().telemetry.addData("Detect Marker", "Detected Marker");
-        robot.getOpmode().telemetry.addData("Vision Status", "Vision initialized");
+        telemetry.telemetry(3, "Detect Marker", "Detected Marker");
+        telemetry.telemetry(2, "Vision Status", "Vision initialized");
+*/
+
+        // Detect marker stuff
+        this.hardwareMap = hardwareMap;
+        this.robotCamera = robotCamera;
+        this.quickTelemetry = quickTelemetry.newQuickTelemetryFile("Detect Marker Pipeline");
+        VisionConfig.finalMarkerLocation = DetectMarkerRun();
     }
 
     private void initVuforia() {
@@ -97,5 +127,38 @@ public class Vision {
     // Helper method to create matrix to identify locations
     public OpenGLMatrix createMatrix(float x, float y, float z, float u, float v, float w) {
         return OpenGLMatrix.translation(x, y, z).multiplied(Orientation.getRotationMatrix(EXTRINSIC, XYZ, DEGREES, u, v, w));
+    }
+
+    /**
+     * This method waits until the search for the marker is done, and then it return the marker
+     * location. It waits until the marker is found, then it returns the marker location.
+     *
+     * @return Where the marker is
+     * @see org.firstinspires.ftc.teamcode.Enhancement.Subsystems.Vision.DetectMarkerPipeline#getMarkerLocation()
+     */
+    public MarkerLocation DetectMarkerRun() {
+        org.firstinspires.ftc.teamcode.Enhancement.Subsystems.Vision.DetectMarkerPipeline detectMarkerPipeline = new DetectMarkerPipeline(quickTelemetry);
+        robotCamera.setPipeline(detectMarkerPipeline);
+
+        // We set the viewport policy to optimized view so the preview doesn't appear 90 deg
+        // out when the RC activity is in portrait. We do our actual image processing assuming
+        // landscape orientation, though.
+        robotCamera.setViewportRenderingPolicy(OpenCvCamera.ViewportRenderingPolicy.OPTIMIZE_VIEW);
+
+        robotCamera.openCameraDeviceAsync(new OpenCvCamera.AsyncCameraOpenListener() {
+            @Override
+            public void onOpened() {
+                robotCamera.startStreaming(CAMERA_WIDTH, CAMERA_HEIGHT, OpenCvCameraRotation.UPRIGHT);
+            }
+
+            @Override
+            public void onError(int errorCode) {
+                markerLocation = MarkerLocation.NOT_FOUND;
+            }
+        });
+        if (!(markerLocation == MarkerLocation.NOT_FOUND)) {
+            markerLocation = detectMarkerPipeline.getMarkerLocation();
+        }
+        return markerLocation;
     }
 }
